repository

+++ src/datas/repositories/EntryRepository.java
@@ -1,6 +1,5 @@
 package datas.repositories;
 
-import datas.models.Diary;
 import datas.models.Entry;
 
 import java.util.List;
@@ -9,7 +8,7 @@ public interface EntryRepository {
 
     Entry save(Entry entry);
     List<Entry> findAll();
-    Entry findById(int id);
+    Entry findById(long id);
     long count();
     void delete(int id);
     void delete(Entry entry);
